refractor: move types to interfaces.ts
